# openapi/internal/scan/v1.yaml
openapi: 3.0.3
info:
  title: CAFE Persistence — internal scan API
  version: "1.0.0"
  license:
    name: Proprietary
    url: https://example.com/license
  description: |
    **PERS-D3a-spec** — service-to-service contract for scan data-plane operations in
    `cafe-persistence`. Complements the async NATS lifecycle (`scan.started`, `scan.completed`,
    `scan.failed`); does **not** replace NATS for scanner execution.

    **Consumers:** `cafe-discovery` API (D6a-read / D6a-delete / D6a-pending).

    **Non-goals (this milestone):** HTTP handlers, edge NGINX routes, module CP (`internal/cp/v1`),
    changes to public `/api/discovery/v1` (`discovery-v1.yaml`).

    **URL prefix (in-cluster):** `/internal/scan/v1` on the persistence internal HTTP listener
    (`PERSISTENCE_INTERNAL_HTTP_PORT`, default `8082`). **Not** exposed on the public product edge.
    Health/readiness probes remain on `PERSISTENCE_HEALTH_PORT` (`/health`, `/ready`).

    **Auth:** static platform bearer token (`CAFE_PERSISTENCE_SERVICE_TOKEN`). Callers derive
    `user_id` / `tenant_id` from the end-user JWT and propagate them via headers — **never** from
    an untrusted client body. See ADR persistence §9.1.

    **Status vocabulary:** persistence-native lifecycle values (`PENDING`, `RUNNING`, `SUCCESS`,
    `FAILED`, `TIMEOUT`, `UNREACHABLE`, empty). Discovery maps these to public v1
    (`requested` / `started` / `completed` / `failed`).
servers:
...
    variables:
      internal_port:
...
security:
  - serviceBearer: []
paths:
  /pending/wallet:
    post:
      tags: [pending]
      summary: Reserve wallet pending scan (atomic per owner+address)
      description: |
        Mirrors `PendingV1ScanRepository.PutWallet` (Redis `SET NX` on
        `discovery:v1:pending_wallet:{user_id}:{address}` plus `discovery:v1:pending_scan:{scan_id}`).

        **409** when another wallet scan is already reserved for the same owner+address (W8).
        Discovery rolls back the reservation when NATS publish fails after a successful reserve.
      operationId: reserveWalletPending
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveWalletPendingRequest"
      responses:
        "201":
          description: Reservation created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservePendingResponse"
        "409":
          description: Wallet scan already in progress for this owner+address.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: SCAN_IN_PROGRESS
                message: A wallet scan is already in progress for this target.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    get:
      tags: [pending]
      summary: Get wallet pending reservation by owner address
      description: |
        Mirrors `GetWalletByOwnerAddress`. Returns the pending record when the address key or
        correlated `pending_scan` payload exists. Used by Discovery W8 in-flight checks.
      operationId: getWalletPendingByAddress
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
        - name: address
          in: query
          required: true
          schema:
            type: string
          description: Normalized Ethereum address (lowercase hex).
      responses:
        "200":
          description: Pending record (may be address-key only when scan payload expired).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingScanRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      tags: [pending]
      summary: Release wallet address reservation (compare-and-delete)
      description: |
        Mirrors `DeleteWalletReservation` — deletes the address key only when its value equals
        `scan_id` (Redis Lua compare-and-delete).
      operationId: releaseWalletPendingReservation
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
        - name: address
          in: query
          required: true
          schema:
            type: string
        - name: scan_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Reservation released (or key absent / mismatched — idempotent no-op).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /pending/tls:
    post:
      tags: [pending]
      summary: Store TLS pending scan correlation
      description: |
        Mirrors `PendingV1ScanRepository.Put` for TLS family (no per-endpoint NX reservation).
      operationId: putTlsPending
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutTlsPendingRequest"
      responses:
        "201":
          description: Pending record stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingScanRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /pending/{scan_id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [pending]
      summary: Get pending scan by scan_id
      operationId: getPendingScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: Pending record for scan_id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingScanRecord"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [pending]
      summary: Release pending scan correlation
      description: |
        Mirrors `Delete` — removes `discovery:v1:pending_scan:{scan_id}` and, for wallet family,
        the correlated address reservation when present.
      operationId: releasePendingScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "204":
          description: Pending keys removed (idempotent).
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /wallets/scans:
    get:
      tags: [wallet-scans]
      summary: List owner wallet scans (persistence rows)
      description: |
        Owner-scoped Postgres list with the same filter semantics as public
        `GET /api/discovery/v1/wallets/scans`. Returns **persistence-native** rows; Discovery
        maps list items to public `ScanListItem` lifecycle synopsis.
      operationId: listWalletScans
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: address
          in: query
          schema:
            type: string
          description: Filter by normalized Ethereum address.
        - name: chain_id
          in: query
          schema:
            type: integer
            format: int64
          description: Requires address query param; filters rows whose networks include the chain.
        - name: latest
          in: query
          schema:
            type: boolean
          description: When true with address, return at most one newest completed-success row.
      responses:
        "200":
          description: Paginated wallet scan rows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedWalletScanRows"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /wallets/scans/{scan_id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [wallet-scans]
      summary: Get owner wallet scan row by scan_id
      operationId: getWalletScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: Wallet scan persistence row (404 when not owner-scoped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletScanRow"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [wallet-scans]
      summary: Delete owner wallet scan row
      description: |
        Soft-deletes the Postgres row and clears correlated Redis cache keys when this was the
        last row for the address. **Does not** perform W3 policy reference checks — Discovery
        guards immutability before calling this endpoint (ADR §9.3).
      operationId: deleteWalletScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "204":
          description: Row deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tls/scans:
    get:
      tags: [tls-scans]
      summary: List owner TLS scans (non-default rows)
      operationId: listTlsScans
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated owner TLS scan rows (default=false).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTlsScanRows"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tls/scans/defaults:
    get:
      tags: [tls-scans]
      summary: List shared default TLS catalog rows
      operationId: listTlsDefaultScans
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: Shared catalog (default=true), not owner-scoped.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTlsScanRows"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /tls/scans/{scan_id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [tls-scans]
      summary: Get TLS scan row by scan_id
      description: |
        Resolves owner rows and shared catalog defaults (`default=true`) by `scan_id`, matching
        public Discovery GET semantics.
      operationId: getTlsScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "200":
          description: TLS scan persistence row.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TlsScanRow"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [tls-scans]
      summary: Delete owner TLS scan row
      description: |
        Soft-deletes an owner TLS row (`default=false`). **Does not** perform W3 policy checks.
      operationId: deleteTlsScan
      parameters:
        - $ref: "#/components/parameters/UserIdHeader"
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/RequestIdHeader"
      responses:
        "204":
          description: Row deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/ServiceUnauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    serviceBearer:
      type: http
      scheme: bearer
      description: |
        Platform service token (`CAFE_PERSISTENCE_SERVICE_TOKEN`). Constant-time comparison
        on the server. Temporary placeholder until mTLS or signed service JWT (ADR §9.1).
  parameters:
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: End-user owner id derived from the caller's validated JWT.
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: false
      schema:
        type: string
      description: Optional tenant scope propagated from the caller.
    RequestIdHeader:
      name: X-Request-Id
      in: header
      required: false
      schema:
        type: string
      description: Correlation id echoed on error responses when provided.
...
  schemas:
...
    ReservePendingResponse:
      type: object
      required: [reserved, record]
      properties:
        reserved:
          type: boolean
          description: True when the address reservation was created (SET NX succeeded).
        record:
          $ref: "#/components/schemas/PendingScanRecord"
...